One signing script: sign.py serve on the YubiKey machine, sign.py sign in the release job
