+++ desktop/src-tauri/tauri.windows.signing.conf.json
@@ -3,7 +3,7 @@
 		"windows": {
 			"signCommand": {
 				"cmd": "uv.exe",
-				"args": ["run", "../../scripts/sign_windows.py", "%1"]
+				"args": ["run", "../../scripts/sign.py", "sign", "%1"]
 			}
 		}
 	}
